+++ src/gui/ClienteFrame.java
@@ -1,10 +1,11 @@
+package gui;
 import javax.swing.*;
 import javax.swing.border.Border;
 import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
-public class NavioFrame extends JFrame implements ActionListener{
+public class ClienteFrame extends JFrame implements ActionListener {
 
     private JPanel nomeEmpresa;
     private int xBotao;
@@ -14,28 +15,48 @@ public class NavioFrame extends JFrame implements ActionListener{
     private Color colorOfButton;
     private Color colorOfLetter;
     private Font fonte;
-    private JPanel nome;
-    private JTextField nomeField;
-    private JPanel velocidade;
-    private JTextField velocidadeField;
-    private JPanel autonomia;
-    private JTextField autonomiaField;
-    private JPanel CustoMilha;
-    private JTextField CustoMilhaField; 
     private JButton selectButton;
     private JButton voltarMenu;
+    //TextField
+    private JTextField codField;
+    private JTextField nomeField;
+    private JTextField emailField;
+
+    private JPanel cod;
+    private JPanel email;
+    private JPanel nome;
     private JLabel errorMessage;
 
-    public NavioFrame(){
+
+   
+
+    public ClienteFrame(){
         super();
         colorOfButton = Color.DARK_GRAY;
         colorOfLetter = Color.WHITE;
         fonte=new Font("Serif", Font.BOLD, 20);
+
         xBotao=100;
         yBotao=150;
         widthBotao=300;
         heightBotao=40;
+        setNomeEmpresa();
+        setDescricaoPagina();
+        setNomePanel();
+        setCodPanel();
+        setEmailPanel();
+        setButtonSelect();
+        setButtonBack();
+        setErrorMessage();
+        this.add(errorMessage);
+        errorMessage.setVisible(false);
 
+        this.add(nomeEmpresa);
+        this.add(cod);
+        this.add(email);
+        this.add(nome);
+        this.add(selectButton);
+        this.add(voltarMenu);
         this.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
         this.setSize(500, 600);
         this.setLayout(null);
@@ -43,45 +64,15 @@ public class NavioFrame extends JFrame implements ActionListener{
         this.setLocationRelativeTo(null);
         this.setResizable(false);
 
-        setDescricaoPagina();
-        setButtonBack();
-        this.add(voltarMenu);
-        setButtonSelect();
-        this.add(selectButton);
-        setNomeEmpresa();
-        this.add(nomeEmpresa);
-        setNomePanel();
-        this.add(nome);
-        setVelocidadePanel();
-        this.add(velocidade);
-        setAutonomiaPanel();
-        this.add(autonomia);
-        setCustoMilhaPanel();
-        this.add(CustoMilha);
-        setErrorMessage();
-        this.add(errorMessage);
-        errorMessage.setVisible(false);
-        
     }
     private void setDescricaoPagina(){
         JPanel descricaoPagina = new JPanel();
-        JLabel label = new JLabel("Cadastrar Navio");
-        label.setFont(new Font ("",Font.BOLD, 20));
+        JLabel label = new JLabel("Cadastrar Cliente");
         label.setForeground(colorOfButton);
+        label.setFont(new Font ("",Font.BOLD, 20));
         descricaoPagina.add(label);
         descricaoPagina.setBounds(0, 100, 500, 40);
         this.add(descricaoPagina);
-    }
-    private void setNomeEmpresa(){
-        //Panel para ter o nome 
-        nomeEmpresa=new JPanel();
-        nomeEmpresa.setBounds(0, 10, 500, 40);
-        JLabel label= new JLabel("Empresa ACME");
-        label.setFont(fonte);
-        label.setForeground(colorOfLetter);
-        nomeEmpresa.add(label);
-        nomeEmpresa.setBackground(colorOfButton);
-        
     }
     private void setNomePanel(){
         nome=new JPanel();
@@ -94,41 +85,36 @@ public class NavioFrame extends JFrame implements ActionListener{
         nome.add(nomeField);
         
     }
-    private void setVelocidadePanel(){
-        velocidade=new JPanel();
-        JLabel label= new JLabel("Velocidade:");
+    private void setCodPanel(){
+
+        JLabel label= new JLabel("Codigo:");
         label.setForeground(colorOfLetter);
-        velocidadeField= new JTextField(11);
-        velocidade.setBackground(colorOfButton);
-        velocidade.setBounds(xBotao,yBotao+75,widthBotao,heightBotao);
-        velocidade.add(label);
-        velocidade.add(velocidadeField);
+        codField= new JTextField(11);
+        cod=new JPanel();
+        cod.setBounds(xBotao,yBotao+75,widthBotao,heightBotao);
+        cod.setBackground(colorOfButton);
+        cod.add(label);
+        cod.add(codField);
+
+
+    }
+    private void setEmailPanel(){
+
+        email=new JPanel();
+        email.setBounds(xBotao,yBotao+150,widthBotao,heightBotao);
+        JLabel label= new JLabel("Email:");
+        label.setForeground(colorOfLetter);
+        emailField= new JTextField(11);
+        email.setBackground(colorOfButton);
+        email.add(label);
+        email.add(emailField);
         
-    }
-    private void setAutonomiaPanel(){
-        autonomia=new JPanel();
-        JLabel label= new JLabel("Autonomia:");
-        label.setForeground(colorOfLetter);
-        autonomiaField= new JTextField(11);
-        autonomia.setBackground(colorOfButton);
-        autonomia.setBounds(xBotao,yBotao+150,widthBotao,heightBotao);
-        autonomia.add(label);
-        autonomia.add(autonomiaField);
-    }
-    private void setCustoMilhaPanel(){
-        CustoMilha=new JPanel();
-        JLabel label= new JLabel("Custo por milha:");
-        label.setForeground(colorOfLetter);
-        CustoMilhaField= new JTextField(11);
-        CustoMilha.setBackground(colorOfButton);
-        CustoMilha.setBounds(xBotao,yBotao+225,widthBotao,heightBotao);
-        CustoMilha.add(label);
-        CustoMilha.add(CustoMilhaField);
+        
     }
     private void setButtonSelect(){
         //Botao de carregar dados
         selectButton = new JButton("Cadastrar");
-        selectButton.setBounds(150, 445, 200, 40);
+        selectButton.setBounds(150, 400, 200, 40);
         selectButton.setFont(fonte);
         selectButton.setFocusable(false);
         selectButton.setForeground(colorOfLetter);
@@ -147,27 +133,40 @@ public class NavioFrame extends JFrame implements ActionListener{
         voltarMenu.addActionListener(this);
         voltarMenu.setBackground(colorOfButton);
     }
+
+
+    private void setNomeEmpresa(){
+        //Panel para ter o nome 
+        nomeEmpresa=new JPanel();
+        nomeEmpresa.setBounds(0, 10, 500, 40);
+        JLabel label= new JLabel("Empresa ACME");
+        label.setFont(fonte);
+        label.setForeground(colorOfLetter);
+        nomeEmpresa.add(label);
+        nomeEmpresa.setBackground(colorOfButton);
+        
+    }
     private void setErrorMessage(){
         errorMessage = new JLabel("Erro ao Cadastrar");
+        errorMessage.setBackground(Color.BLACK);
         errorMessage.setForeground(Color.RED);
         errorMessage.setFont(new Font("Serif", Font.BOLD, 15));
         errorMessage.setBounds(175, 300, 500, 400);
     }
 
-
     @Override
     public void actionPerformed(ActionEvent e) {
         if(e.getSource() == selectButton){
             try{
-                String n =nomeField.getText();
-                double v= Double.parseDouble(velocidadeField.getText());
-                double auto = Double.parseDouble(autonomiaField.getText());
-                double custo = Double.parseDouble(CustoMilhaField.getText());
-                //Navio n = new Navio(n,v,auto,custo);
-                //ColecaoNavio.add(n);
+                int cod = Integer.parseInt(codField.getText());
+                String nome = nomeField.getText();
+                String email = nomeField.getText();
+                //Cliente novoCliente = new Cliente(cod,nome,email);
+
+                //Adicionar CLiente na colecao
                 errorMessage.setVisible(false);
-                
-            }catch(NumberFormatException except){
+            }
+            catch(NumberFormatException except){
                 errorMessage.setVisible(true);
             }
         }
@@ -176,4 +175,5 @@ public class NavioFrame extends JFrame implements ActionListener{
             dispose();
         }
     }
+   
 }
