+++ src/gui/DadosFrame.java
@@ -1,3 +1,4 @@
+package gui;
 import javax.swing.*;
 import javax.swing.border.Border;
 import java.awt.*;
