+++ src/gui/PortoFrame.java
@@ -1,11 +1,11 @@
+package gui;
 import javax.swing.*;
 import javax.swing.border.Border;
 import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
-public class ClienteFrame extends JFrame implements ActionListener {
-
+public class PortoFrame extends JFrame implements ActionListener{
     private JPanel nomeEmpresa;
     private int xBotao;
     private int yBotao;
@@ -13,23 +13,24 @@ public class ClienteFrame extends JFrame implements ActionListener {
     private int heightBotao;
     private Color colorOfButton;
     private Color colorOfLetter;
-    private Font fonte;
+    private Font fonte; 
+
+    private JPanel nome;
+    private JPanel id;
+    private JPanel pais;
     private JButton selectButton;
     private JButton voltarMenu;
-    //TextField
-    private JTextField codField;
-    private JTextField nomeField;
-    private JTextField emailField;
-
-    private JPanel cod;
-    private JPanel email;
-    private JPanel nome;
+    private JTextField fieldNome;
+    private JTextField fieldId;
+    private JTextField fieldPais;
     private JLabel errorMessage;
 
+    
+    
 
-   
+    
 
-    public ClienteFrame(){
+    public PortoFrame(){
         super();
         colorOfButton = Color.DARK_GRAY;
         colorOfLetter = Color.WHITE;
@@ -39,21 +40,23 @@ public class ClienteFrame extends JFrame implements ActionListener {
         yBotao=150;
         widthBotao=300;
         heightBotao=40;
+
+
+        setNome();
+        setId();
+        setPais();
         setNomeEmpresa();
-        setDescricaoPagina();
-        setNomePanel();
-        setCodPanel();
-        setEmailPanel();
         setButtonSelect();
+        setDescricaoPagina();
         setButtonBack();
+
         setErrorMessage();
         this.add(errorMessage);
         errorMessage.setVisible(false);
-
-        this.add(nomeEmpresa);
-        this.add(cod);
-        this.add(email);
         this.add(nome);
+        this.add(nomeEmpresa);
+        this.add(id);
+        this.add(pais);
         this.add(selectButton);
         this.add(voltarMenu);
         this.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
@@ -63,50 +66,61 @@ public class ClienteFrame extends JFrame implements ActionListener {
         this.setLocationRelativeTo(null);
         this.setResizable(false);
 
+
+        
     }
     private void setDescricaoPagina(){
         JPanel descricaoPagina = new JPanel();
-        JLabel label = new JLabel("Cadastrar Cliente");
+        JLabel label = new JLabel("Cadastrar Porto");
         label.setForeground(colorOfButton);
         label.setFont(new Font ("",Font.BOLD, 20));
         descricaoPagina.add(label);
         descricaoPagina.setBounds(0, 100, 500, 40);
         this.add(descricaoPagina);
+
     }
-    private void setNomePanel(){
+    private void setNomeEmpresa(){
+        //Panel para ter o nome 
+        nomeEmpresa=new JPanel();
+        nomeEmpresa.setBounds(0, 10, 500, 40);
+        JLabel label= new JLabel("Empresa ACME");
+        label.setFont(fonte);
+        label.setForeground(colorOfLetter);
+        nomeEmpresa.add(label);
+        nomeEmpresa.setBackground(colorOfButton);
+        
+    }
+    private void setNome(){
         nome=new JPanel();
         JLabel label= new JLabel("Nome:");
         label.setForeground(colorOfLetter);
-        nomeField= new JTextField(11);
+        fieldNome= new JTextField(11);
         nome.setBackground(colorOfButton);
         nome.setBounds(xBotao,yBotao,widthBotao,heightBotao);
         nome.add(label);
-        nome.add(nomeField);
+        nome.add(fieldNome);
         
     }
-    private void setCodPanel(){
+    private void setId(){
 
-        JLabel label= new JLabel("Codigo:");
+        JLabel label= new JLabel("ID:");
         label.setForeground(colorOfLetter);
-        codField= new JTextField(11);
-        cod=new JPanel();
-        cod.setBounds(xBotao,yBotao+75,widthBotao,heightBotao);
-        cod.setBackground(colorOfButton);
-        cod.add(label);
-        cod.add(codField);
-
-
+        fieldId= new JTextField(11);
+        id=new JPanel();
+        id.setBounds(xBotao,yBotao+75,widthBotao,heightBotao);
+        id.setBackground(colorOfButton);
+        id.add(label);
+        id.add(fieldId);
     }
-    private void setEmailPanel(){
-
-        email=new JPanel();
-        email.setBounds(xBotao,yBotao+150,widthBotao,heightBotao);
-        JLabel label= new JLabel("Email:");
+    private void setPais(){
+        pais=new JPanel();
+        pais.setBounds(xBotao,yBotao+150,widthBotao,heightBotao);
+        JLabel label= new JLabel("País:");
         label.setForeground(colorOfLetter);
-        emailField= new JTextField(11);
-        email.setBackground(colorOfButton);
-        email.add(label);
-        email.add(emailField);
+        fieldPais= new JTextField(11);
+        pais.setBackground(colorOfButton);
+        pais.add(label);
+        pais.add(fieldPais);
         
         
     }
@@ -132,19 +146,6 @@ public class ClienteFrame extends JFrame implements ActionListener {
         voltarMenu.addActionListener(this);
         voltarMenu.setBackground(colorOfButton);
     }
-
-
-    private void setNomeEmpresa(){
-        //Panel para ter o nome 
-        nomeEmpresa=new JPanel();
-        nomeEmpresa.setBounds(0, 10, 500, 40);
-        JLabel label= new JLabel("Empresa ACME");
-        label.setFont(fonte);
-        label.setForeground(colorOfLetter);
-        nomeEmpresa.add(label);
-        nomeEmpresa.setBackground(colorOfButton);
-        
-    }
     private void setErrorMessage(){
         errorMessage = new JLabel("Erro ao Cadastrar");
         errorMessage.setBackground(Color.BLACK);
@@ -157,20 +158,21 @@ public class ClienteFrame extends JFrame implements ActionListener {
     public void actionPerformed(ActionEvent e) {
         if(e.getSource() == selectButton){
             try{
-                int cod = Integer.parseInt(codField.getText());
-                String nome = nomeField.getText();
-                String email = nomeField.getText();
-                //Cliente novoCliente = new Cliente(cod,nome,email);
+                String nome = fieldNome.getText();
+                int id =Integer.parseInt(fieldId.getText());
+                String pais = fieldPais.getText();
+                //Porto p = new Porto(nome,id,pais);
+                //Adicionar o porto na colecao
                 errorMessage.setVisible(false);
-            }
-            catch(NumberFormatException except){
+            }catch(NumberFormatException except){
                 errorMessage.setVisible(true);
             }
+
         }
         if(e.getSource() == voltarMenu){
             MenuInicial menu = new MenuInicial();
             dispose();
         }
     }
-   
+    
 }
